+++ src/components/restaurants/LeftNavBar.styled.ts
@@ -1,0 +1,9 @@
+import { styled } from '@mui/styles';
+import { Box } from '@mui/material';
+
+export const StyledLayout = styled(Box)({
+  width: '262px',
+  height: '100vh',
+  marginRight: '24px',
+  background: '#ccc',
+});